add camera screen and nav bar icon
