+++ navigation/MainTabNavigator.js
@@ -4,8 +4,8 @@ import { createStackNavigator, createBottomTabNavigator } from 'react-navigation
 
 import TabBarIcon from '../components/TabBarIcon';
 import HomeScreen from '../screens/HomeScreen';
-import LinksScreen from '../screens/LinksScreen';
 import SettingsScreen from '../screens/SettingsScreen';
+import CameraScreen from '../screens/CameraScreen';
 
 const config = Platform.select({
   web: { headerMode: 'screen' },
@@ -28,11 +28,27 @@ HomeStack.navigationOptions = {
     />
   ),
   tabBarOptions: {
-    style: {backgroundColor: 'white', borderTopWidth: 0, shadowColor: 'black', shadowOffset:{width: 5, height: 3}, shadowOpacity: 0.5, elevation: 10}
+    style: { backgroundColor: 'white', borderTopWidth: 0, shadowColor: 'black', shadowOffset: { width: 5, height: 3 }, shadowOpacity: 0.5, elevation: 10 }
   }
 };
 
 HomeStack.path = '';
+
+const CameraStack = createStackNavigator(
+  {
+    Camera: CameraScreen,
+  },
+  config
+);
+
+CameraStack.navigationOptions = {
+  tabBarLabel: 'Camera',
+  tabBarIcon: ({ focused }) => (
+    <TabBarIcon focused={focused} name={'ios-camera'} />
+  ),
+}
+
+CameraStack.path = '';
 
 const SettingsStack = createStackNavigator(
   {
@@ -53,6 +69,7 @@ SettingsStack.path = '';
 // the nav icons at the bottom
 const tabNavigator = createBottomTabNavigator({
   HomeStack,
+  CameraStack,
   SettingsStack,
 });
